learn: #62 - 엔티티 클래스와 JPA 매핑 정보 분리
