--- spring-lab/splearn/src/main/java/org/example/splean/SpleanApplication.java
@@ -1,13 +1,0 @@
-package org.example.splean;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class SpleanApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(SpleanApplication.class, args);
-    }
-
-}
