--- spring-lab/splearn/src/main/java/org/example/splean/adapter/integration/DummyEmailSender.java
@@ -1,15 +1,0 @@
-package org.example.splean.adapter.integration;
-
-import org.example.splean.application.required.EmailSender;
-import org.example.splean.domain.Email;
-import org.springframework.context.annotation.Fallback;
-import org.springframework.stereotype.Component;
-
-@Component
-@Fallback
-public class DummyEmailSender implements EmailSender {
-    @Override
-    public void send(Email email, String subject, String body) {
-        System.out.println("DummyEmailSender send email: " + email);
-    }
-}
