--- spring-lab/splearn/src/main/java/org/example/splean/application/required/EmailSender.java
@@ -1,10 +1,0 @@
-package org.example.splean.application.required;
-
-import org.example.splean.domain.Email;
-
-/**
- * 이메일을 발송한다
- */
-public interface EmailSender {
-    void send(Email email, String subject, String body);
-}
