--- spring-lab/splearn/src/main/java/org/example/splean/application/required/MemberRepository.java
@@ -1,18 +1,0 @@
-package org.example.splean.application.required;
-
-import org.example.splean.domain.Email;
-import org.example.splean.domain.Member;
-import org.springframework.data.repository.Repository;
-
-import java.util.Optional;
-
-/**
- * 회원 정보를 저장하거나 조회한다
- */
-public interface MemberRepository extends Repository<Member, Long> {
-    Member save(Member member);
-
-    Optional<Member> findByEmail(Email email);
-
-    Optional<Member> findById(Long memberId);
-}
