--- spring-lab/splearn/src/main/java/org/example/splean/domain/DuplicateEmailException.java
@@ -1,6 +1,0 @@
-package org.example.splean.domain;
-
-public class DuplicateEmailException extends RuntimeException {
-    public DuplicateEmailException(String message) {
-    }
-}
