--- spring-lab/splearn/src/main/java/org/example/splean/domain/Email.java
@@ -1,19 +1,0 @@
-package org.example.splean.domain;
-
-import jakarta.persistence.Embeddable;
-
-import java.util.regex.Pattern;
-
-@Embeddable
-public record Email(String address) {
-
-    private static final Pattern EMAIL_PATTERN = Pattern.compile(
-            "^[a-zA-Z0-9_+&*-]+(?:\\.[a-zA-Z0-9_+&*-]+)*@(?:[a-zA-Z0-9-]+\\.)+[a-zA-Z]{2,7}$");
-
-
-    public Email {
-        if (!Email.EMAIL_PATTERN.matcher(address).matches()) {
-            throw new IllegalArgumentException("이메일 형식이 바르지 않습니다." + address);
-        }
-    }
-}
