--- spring-lab/splearn/src/main/java/org/example/splean/domain/MemberRegisterRequest.java
@@ -1,11 +1,0 @@
-package org.example.splean.domain;
-
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.Size;
-
-public record MemberRegisterRequest(
-        @Email String email,
-        @Size(min = 5, max = 20) String nickname,
-        @Size(min = 8, max = 80) String password) {
-
-}
