--- spring-lab/splearn/src/main/java/org/example/splean/domain/MemberStatus.java
@@ -1,5 +1,0 @@
-package org.example.splean.domain;
-
-public enum MemberStatus {
-    PENDING, ACTIVE, DEACTIVATED
-}
