--- spring-lab/splearn/src/main/java/org/example/splean/domain/PasswordEncoder.java
@@ -1,7 +1,0 @@
-package org.example.splean.domain;
-
-public interface PasswordEncoder {
-    String encode(String password);
-
-    boolean matches(String password, String passwordHash);
-}
