--- spring-lab/splearn/src/main/java/org/example/splean/domain/package-info.java
@@ -1,4 +1,0 @@
-@NonNullApi
-package org.example.splean.domain;
-
-import org.springframework.lang.NonNullApi;
