+++ spring-lab/splearn/src/main/java/org/example/splearn/SpleanApplication.java
@@ -1,0 +1,13 @@
+package org.example.splearn;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SpleanApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(SpleanApplication.class, args);
+    }
+
+}
