+++ spring-lab/splearn/src/main/java/org/example/splearn/adapter/integration/DummyEmailSender.java
@@ -1,0 +1,15 @@
+package org.example.splearn.adapter.integration;
+
+import org.example.splearn.application.required.EmailSender;
+import org.example.splearn.domain.Email;
+import org.springframework.context.annotation.Fallback;
+import org.springframework.stereotype.Component;
+
+@Component
+@Fallback
+public class DummyEmailSender implements EmailSender {
+    @Override
+    public void send(Email email, String subject, String body) {
+        System.out.println("DummyEmailSender send email: " + email);
+    }
+}
