+++ spring-lab/splearn/src/main/java/org/example/splearn/adapter/security/SecurePasswordEncoder.java
@@ -1,0 +1,20 @@
+package org.example.splearn.adapter.security;
+
+import org.example.splearn.domain.PasswordEncoder;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.stereotype.Component;
+
+@Component
+public class SecurePasswordEncoder implements PasswordEncoder {
+    private final BCryptPasswordEncoder byCryptPasswordEncoder = new BCryptPasswordEncoder();
+
+    @Override
+    public String encode(String password) {
+        return byCryptPasswordEncoder.encode(password);
+    }
+
+    @Override
+    public boolean matches(String password, String passwordHash) {
+        return byCryptPasswordEncoder.matches(password, passwordHash);
+    }
+}
