+++ spring-lab/splearn/src/main/java/org/example/splearn/application/MemberModifyService.java
@@ -1,0 +1,55 @@
+package org.example.splearn.application;
+
+import lombok.RequiredArgsConstructor;
+import org.example.splearn.application.provided.MemberFinder;
+import org.example.splearn.application.provided.MemberRegister;
+import org.example.splearn.application.required.EmailSender;
+import org.example.splearn.application.required.MemberRepository;
+import org.example.splearn.domain.*;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.validation.annotation.Validated;
+
+@Validated
+@Service
+@RequiredArgsConstructor
+public class MemberModifyService implements MemberRegister {
+    private final MemberFinder memberFinder;
+    private final MemberRepository memberRepository;
+    private final EmailSender emailSender;
+    private final PasswordEncoder passwordEncoder;
+
+    @Transactional
+    @Override
+    public Member register(MemberRegisterRequest registerRequest) {
+        checkDuplicateEmail(registerRequest);
+
+        Member member = Member.register(registerRequest, passwordEncoder);
+
+        memberRepository.save(member);
+
+        sendWelcomeEmail(member);
+
+        return member;
+    }
+
+    @Transactional
+    @Override
+    public Member activate(Long memberId) {
+        Member member = memberFinder.find(memberId);
+        member.activate();
+
+        return memberRepository.save(member);
+    }
+
+
+    private void sendWelcomeEmail(Member member) {
+        emailSender.send(member.getEmail(), "등록을 완료해주세요", "아래 링크를 클릭해서 등록을 완료해주세요");
+    }
+
+    private void checkDuplicateEmail(MemberRegisterRequest registerRequest) {
+        if (memberRepository.findByEmail(new Email(registerRequest.email())).isPresent()) {
+            throw new DuplicateEmailException("이미 등록된 이메일입니다: " + registerRequest.email());
+        }
+    }
+}
