+++ spring-lab/splearn/src/main/java/org/example/splearn/domain/DuplicateEmailException.java
@@ -1,0 +1,6 @@
+package org.example.splearn.domain;
+
+public class DuplicateEmailException extends RuntimeException {
+    public DuplicateEmailException(String message) {
+    }
+}
