+++ spring-lab/splearn/src/main/java/org/example/splearn/domain/Email.java
@@ -1,0 +1,16 @@
+package org.example.splearn.domain;
+
+import java.util.regex.Pattern;
+
+public record Email(String address) {
+
+    private static final Pattern EMAIL_PATTERN = Pattern.compile(
+            "^[a-zA-Z0-9_+&*-]+(?:\\.[a-zA-Z0-9_+&*-]+)*@(?:[a-zA-Z0-9-]+\\.)+[a-zA-Z]{2,7}$");
+
+
+    public Email {
+        if (!Email.EMAIL_PATTERN.matcher(address).matches()) {
+            throw new IllegalArgumentException("이메일 형식이 바르지 않습니다." + address);
+        }
+    }
+}
