+++ spring-lab/splearn/src/main/java/org/example/splearn/domain/MemberRegisterRequest.java
@@ -1,0 +1,11 @@
+package org.example.splearn.domain;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.Size;
+
+public record MemberRegisterRequest(
+        @Email String email,
+        @Size(min = 5, max = 20) String nickname,
+        @Size(min = 8, max = 80) String password) {
+
+}
