+++ spring-lab/splearn/src/main/java/org/example/splearn/domain/MemberStatus.java
@@ -1,0 +1,5 @@
+package org.example.splearn.domain;
+
+public enum MemberStatus {
+    PENDING, ACTIVE, DEACTIVATED
+}
