+++ spring-lab/splearn/src/main/java/org/example/splearn/domain/PasswordEncoder.java
@@ -1,0 +1,7 @@
+package org.example.splearn.domain;
+
+public interface PasswordEncoder {
+    String encode(String password);
+
+    boolean matches(String password, String passwordHash);
+}
