+++ spring-lab/splearn/src/main/java/org/example/splearn/domain/package-info.java
@@ -1,0 +1,4 @@
+@NonNullApi
+package org.example.splearn.domain;
+
+import org.springframework.lang.NonNullApi;
