--- spring-lab/splearn/src/test/java/org/example/splean/adapter/integration/DummyEmailSenderTest.java
@@ -1,20 +1,0 @@
-package org.example.splean.adapter.integration;
-
-import org.example.splean.domain.Email;
-import org.junit.jupiter.api.Test;
-import org.junitpioneer.jupiter.StdIo;
-import org.junitpioneer.jupiter.StdOut;
-
-import static org.assertj.core.api.Assertions.assertThat;
-
-class DummyEmailSenderTest {
-    @Test
-    @StdIo
-    void dummyEmailSender(StdOut out) {
-        DummyEmailSender dummyEmailSender = new DummyEmailSender();
-        dummyEmailSender.send(new Email("ckaanf@gmail.com"), "subject", "body");
-
-        assertThat(out.capturedLines()[0]).isEqualTo("DummyEmailSender send email: Email[address=ckaanf@gmail.com]");
-    }
-
-}
