--- spring-lab/splearn/src/test/java/org/example/splean/application/provided/MemberFinderTest.java
@@ -1,36 +1,0 @@
-package org.example.splean.application.provided;
-
-import jakarta.persistence.EntityManager;
-import org.example.splean.SplearnTestConfiguration;
-import org.example.splean.domain.Member;
-import org.example.splean.domain.MemberFixture;
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.context.annotation.Import;
-import org.springframework.transaction.annotation.Transactional;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-
-@SpringBootTest
-@Transactional
-@Import(SplearnTestConfiguration.class)
-record MemberFinderTest(MemberFinder memberFinder, MemberRegister memberRegister, EntityManager entityManager) {
-
-    @Test
-    void find() {
-        Member member = memberRegister.register(MemberFixture.createMemberRegisterRequest());
-        entityManager.flush();
-        entityManager.clear();
-
-        Member foundMember = memberFinder.find(member.getId());
-
-        assertThat(member.getId()).isEqualTo(foundMember.getId());
-    }
-
-    @Test
-    void findFail() {
-        assertThatThrownBy(() -> memberFinder.find(999L))
-                .isInstanceOf(IllegalArgumentException.class);
-    }
-}
