--- spring-lab/splearn/src/test/java/org/example/splean/application/required/MemberRepositoryTest.java
@@ -1,47 +1,0 @@
-package org.example.splean.application.required;
-
-import jakarta.persistence.EntityManager;
-import org.example.splean.domain.Member;
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.orm.jpa.DataJpaTest;
-import org.springframework.dao.DataIntegrityViolationException;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.example.splean.domain.MemberFixture.createMemberRegisterRequest;
-import static org.example.splean.domain.MemberFixture.createPasswordEncoder;
-
-@DataJpaTest
-class MemberRepositoryTest {
-    @Autowired
-    MemberRepository memberRepository;
-
-    @Autowired
-    EntityManager entityManager;
-
-    @Test
-    void createMember() {
-        Member member = Member.register(createMemberRegisterRequest(), createPasswordEncoder());
-
-        assertThat(member.getId()).isNull();
-
-        memberRepository.save(member);
-
-        assertThat(member.getId()).isNotNull();
-
-        entityManager.flush();
-    }
-
-    @Test
-    void duplicatedEmailFail() {
-        Member member = Member.register(createMemberRegisterRequest(), createPasswordEncoder());
-        memberRepository.save(member);
-
-        Member member2 = Member.register(createMemberRegisterRequest(), createPasswordEncoder());
-
-        assertThatThrownBy(() -> memberRepository.save(member2))
-                .isInstanceOf(DataIntegrityViolationException.class);
-    }
-
-}
