+++ spring-lab/splearn/src/test/java/org/example/splearn/adapter/integration/DummyEmailSenderTest.java
@@ -1,0 +1,20 @@
+package org.example.splearn.adapter.integration;
+
+import org.example.splearn.domain.Email;
+import org.junit.jupiter.api.Test;
+import org.junitpioneer.jupiter.StdIo;
+import org.junitpioneer.jupiter.StdOut;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+class DummyEmailSenderTest {
+    @Test
+    @StdIo
+    void dummyEmailSender(StdOut out) {
+        DummyEmailSender dummyEmailSender = new DummyEmailSender();
+        dummyEmailSender.send(new Email("ckaanf@gmail.com"), "subject", "body");
+
+        assertThat(out.capturedLines()[0]).isEqualTo("DummyEmailSender send email: Email[address=ckaanf@gmail.com]");
+    }
+
+}
