+++ spring-lab/splearn/src/test/java/org/example/splearn/application/provided/MemberFinderTest.java
@@ -1,0 +1,36 @@
+package org.example.splearn.application.provided;
+
+import jakarta.persistence.EntityManager;
+import org.example.splearn.SplearnTestConfiguration;
+import org.example.splearn.domain.Member;
+import org.example.splearn.domain.MemberFixture;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+import org.springframework.transaction.annotation.Transactional;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+
+@SpringBootTest
+@Transactional
+@Import(SplearnTestConfiguration.class)
+record MemberFinderTest(MemberFinder memberFinder, MemberRegister memberRegister, EntityManager entityManager) {
+
+    @Test
+    void find() {
+        Member member = memberRegister.register(MemberFixture.createMemberRegisterRequest());
+        entityManager.flush();
+        entityManager.clear();
+
+        Member foundMember = memberFinder.find(member.getId());
+
+        assertThat(member.getId()).isEqualTo(foundMember.getId());
+    }
+
+    @Test
+    void findFail() {
+        assertThatThrownBy(() -> memberFinder.find(999L))
+                .isInstanceOf(IllegalArgumentException.class);
+    }
+}
