+++ spring-lab/splearn/src/test/java/org/example/splearn/domain/MemberFixture.java
@@ -1,0 +1,25 @@
+package org.example.splearn.domain;
+
+public class MemberFixture {
+    public static MemberRegisterRequest createMemberRegisterRequest() {
+        return createMemberRegisterRequest("ckaanf@test.com");
+    }
+
+    public static MemberRegisterRequest createMemberRegisterRequest(String mail) {
+        return new MemberRegisterRequest(mail, "ckaanf", "secretsecret");
+    }
+
+    public static PasswordEncoder createPasswordEncoder() {
+        return new PasswordEncoder() {
+            @Override
+            public String encode(String password) {
+                return password.toUpperCase();
+            }
+
+            @Override
+            public boolean matches(String password, String passwordHash) {
+                return encode(password).equals(passwordHash);
+            }
+        };
+    }
+}
