+++ spring-lab/splearn/src/test/java/org/example/splearn/domain/MemberTest.java
@@ -1,0 +1,107 @@
+package org.example.splearn.domain;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.assertj.core.api.Assertions.assertThatThrownBy;
+import static org.example.splearn.domain.MemberFixture.createMemberRegisterRequest;
+import static org.example.splearn.domain.MemberFixture.createPasswordEncoder;
+
+class MemberTest {
+    Member member;
+    PasswordEncoder passwordEncoder;
+
+    @BeforeEach
+    void setUp() {
+        this.passwordEncoder = createPasswordEncoder();
+        member = Member.register(createMemberRegisterRequest("ckaanf@splearn.app"), passwordEncoder);
+    }
+
+    @Test
+    void registerMember() {
+
+        assertThat(member.getStatus()).isEqualTo(MemberStatus.PENDING);
+    }
+
+    @Test
+    void activate() {
+
+        member.activate();
+
+        assertThat(member.getStatus()).isEqualTo(MemberStatus.ACTIVE);
+    }
+
+    @Test
+    void activateFail() {
+
+        member.activate();
+
+        assertThatThrownBy(member::activate).isInstanceOf(IllegalStateException.class);
+
+    }
+
+    @Test
+    void deactivate() {
+
+        member.activate();
+        member.deactivate();
+
+        assertThat(member.getStatus()).isEqualTo(MemberStatus.DEACTIVATED);
+    }
+
+    @Test
+    void deactivateFail() {
+
+        assertThatThrownBy(member::deactivate).isInstanceOf(IllegalStateException.class);
+
+        member.activate();
+        member.deactivate();
+
+        assertThatThrownBy(member::deactivate).isInstanceOf(IllegalStateException.class);
+    }
+
+    @Test
+    void verifyPassword() {
+        assertThat(member.verifyPassword("secretsecret", passwordEncoder)).isTrue();
+        assertThat(member.verifyPassword("hellohellohello", passwordEncoder)).isFalse();
+    }
+
+    @Test
+    void changeNickname() {
+        assertThat(member.getNickname()).isEqualTo("ckaanf");
+
+        member.changeNickname("ckacka");
+
+        assertThat(member.getNickname()).isEqualTo("ckacka");
+    }
+
+    @Test
+    void changePassword() {
+        member.changePassword("newSecret", passwordEncoder);
+
+        assertThat(member.verifyPassword("newSecret", passwordEncoder)).isTrue();
+    }
+
+    @Test
+    void isActive() {
+        assertThat(member.isActive()).isFalse();
+
+        member.activate();
+        assertThat(member.isActive()).isTrue();
+
+        member.deactivate();
+        assertThat(member.isActive()).isFalse();
+    }
+
+    @Test
+    void inValidEmail() {
+        assertThatThrownBy(() ->
+                Member.register(createMemberRegisterRequest("test"), passwordEncoder)
+        ).isInstanceOf(IllegalArgumentException.class);
+
+        Member.register(createMemberRegisterRequest("test_test@gmail.com"), passwordEncoder);
+
+    }
+
+}
